t/xml/pis-2.dat

#data
<!DOCTYPE a[
<?x?>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x y?>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <?x y?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
<?y?>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
]>
<?y?>
<a/>
#errors
#document
| <!DOCTYPE a>
|   <?x ?>
| <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ?>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ??y?>
]><a/>
#errors
#document
| <!DOCTYPE a>
|   <?x ??y?>
| <{} a>

#data
<!DOCTYPE a[
<?x
#errors
2;3;no pic;;;m
2;3;unclosed internal subset;;;m
2;3;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<?x\u0020
#errors
2;4;no pic;;;m
2;4;unclosed internal subset;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x y
#errors
2;5;no pic;;;m
2;5;unclosed internal subset;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x ?
#errors
2;5;no pic;;;m
2;5;unclosed internal subset;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?
#errors
2;1;bare pio;;;m
2;2;unclosed internal subset;;;m
2;2;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
#errors
2;3;no pic;;;m
2;3;unclosed internal subset;;;m
2;3;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
]><a/>
#errors
3;6;no pic;;;m
3;6;unclosed internal subset;;;m
3;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<??>
<?x?>
]><a/>
#errors
2;1;bare pio;;;m
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
3;1;stop processing;;;i
3;2;string in internal subset;;;m
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
4;2;string in internal subset;;;m
#xml-standalone
true
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>
